migration and database fixed
